abstract expiration status highlighting in table
